Register Working + Cookie Set in Front-End
The full path of registering a user works from passing front end form data to the back end. Back end data is validated and saved into the database and returns the newly created user to the front end. The front end then saves the user as a cookie before redirecting to the Home page which now displays the user's username on the main page, if they are logged in.