Edit Account + Account View + Logged In Status Security
Edit Account form made and edit account route made. Properly updates the user's account with the form data. User can also view their account information on the account page. Links updated based on cookie state for the home page as well as rerouting in place for when a user tries to view certain pages by changing the url itself. Logged in users are the only ones that can view the account and edit account pages while a non logged in user is only able to go to the login and register pages.