Song View + URLs
Every search result now has a corresponding button to take the user to the Song page. This page shows some basic information about the song and has two URLs that go to the songs guitar tabs and lyrics. Both URLs open separate tabs from the website for convenience of the user.