Preview Button Bug Fix, Playlist Form, View Playlist, Create Playlist
The preview button now sets the state of every single search result individually upon pressing the preview button. Playlist form has been made with a new page as well as allowing the user to view their current playlists in another page. Pages also have security to prevent non logged in users from accessing the pages. User can also create a new playlist with a title and description. Each playlist is saved with the user's ID in order to get all the playlists that match the ID of the current user ONLY.