EVERYTHING CHANGED....sort of
The Deezer API has been completely replaced in favor of Spotify so that the user can always be provided a Spotify link to listen to the songs that they search up. There isn't as much variety with searching songs now, but it is a lot more consistent. To add on further, the Spotify API is also being used to get links for the user to click on so they can listen to the full track on Spotify. In exchange for not having full mp3 tracks, I decided to add in a brand new feature involving a new API called FreeDP API. This API allows the user to search certain categories to find and then proceed to download Royalty Free Tracks! The catch is that only logged in users have access to this feature. Other UX stuff has been put in place to make the website flow more smoothly.