Reviews + Redirect UX + Cookie Expires
Reviews are now in for the main tracks! A logged in user can leave a review on the song info page for a specific song and that review even has a 5 star rating system! This rating system will have some changes later on to make it fit better with the website, but it works! There was also some UX issues going on with a user not being logged in and trying to submit a review. To fix this issue, I allowed a non-logged in user to attempt to make a review, but tell them to log in afterwards. The site redirects them to the Login page. When the user logs in or registers AFTER being redirected from the song info page, the user will then be redirected BACK to the song info page directly with the cookie created. Speaking of the user cookie, now the cookie has an expiration date. This expiration date is exactly 24 hours from the moment they log in to the website.