Delete Account + File Upload Attempt
The user can now freely delete their account if they choose to do so. Also, file upload was attempted, but the file would be lost when being sent to the back end. Research was done to find only outdated methods that are most likely obsolete at this point. An attempt was made, but it is a little too late to focus too much on adding this small feature.