CSS, Preview Bug Fix, Logo
A bug fix has been applied to the preview audio for the case when search results would not return preview urls for every single result. CSS has begun final development and has mostly been finished for a large part of the website. A logo has been made and implemented into the home page of the website, it shall be formatted into more of the website as the styling is finished up.